Rename header in ReSolve lib, print warning if IR unsuccessful.
